refactor: unit test relocation
