+++ interpreter/binaryChannelReader_test.go
@@ -47,40 +47,6 @@ func TestGetHeaders(t *testing.T) {
 	}
 }
 
-func TestBinaryToText(t *testing.T) {
-	tests := []struct {
-		binaryString string
-		expectedText string
-		expectError  bool
-	}{
-		{"0100100001100101011011000110110001101111", "Hello", false},
-		{"0110100001100101011011000110110001101111", "hello", false},
-		{"010000010100001001000011", "ABC", false},
-		{"001100010011001000110011", "123", false},
-		{"", "", false},             // empty string should return an empty string
-		{"00000000", "\x00", false}, // single null byte
-		{"0100100", "", true},       // invalid length (not a multiple of 8)
-		{"0100100G", "", true},      // invalid character in binary string
-	}
-
-	communicationChannel := make(chan models.Binary, len(validHeadersForTextMessage))
-
-	channelReader := BinaryDataInterpreter{
-		Channel: communicationChannel,
-	}
-
-	for _, test := range tests {
-		result, err := channelReader.binaryToText(test.binaryString)
-		if (err != nil) != test.expectError {
-			t.Errorf("binaryToText(%q) error = %v, expected error = %v", test.binaryString, err, test.expectError)
-			continue
-		}
-		if result != test.expectedText {
-			t.Errorf("binaryToText(%q) = %q, want %q", test.binaryString, result, test.expectedText)
-		}
-	}
-}
-
 func TestValidMessageIsCorrectlyChecked(t *testing.T) {
 	tests := []struct {
 		expectedBinaryChecksum string
